+++ styles/globals.css
@@ -6,11 +6,6 @@ body {
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
-/* main {
-  width: 80vw;
-  margin: 0 auto;
-} */
-
 a {
   color: inherit;
   text-decoration: none;
